+++ scripts/output/validation_cell.R
@@ -1,0 +1,207 @@
+# |  (C) 2008-2024 Potsdam Institute for Climate Impact Research (PIK)
+# |  authors, and contributors see CITATION.cff file. This file is part
+# |  of MAgPIE and licensed under AGPL-3.0-or-later. Under Section 7 of
+# |  AGPL-3.0, you are granted additional permissions described in the
+# |  MAgPIE License Exception, version 1.0 (see LICENSE file).
+# |  Contact: magpie@pik-potsdam.de
+
+# ---------------------------------------------------------------------------------------------------
+# description: Writes a validation file where land use types and main crops (soybean, maize, rice and 
+# temparate cereals/wheat) at cellular level are compared to LUH and SPAM initial values.
+# ---------------------------------------------------------------------------------------------------
+
+library(magpie4)
+library(luscale)
+library(mrcommons)
+library(luplot)
+library(mrmagpie)
+library(lusweave)
+library(magpiesets)
+
+############################# BASIC CONFIGURATION ##############################
+map_file <- Sys.glob(file.path(outputdir, "/clustermap_*.rds"))
+gdx <- file.path(outputdir, "/fulldata.gdx")
+cfg <- gms::loadConfig(file.path(outputdir, "/config.yml"))
+crops<-c("maiz","rice_pro","soybean","tece")
+################################################################################
+
+#### Folder for saving results
+out_dir <- paste0(outputdir, "/CellularValidation/")
+if (!dir.exists(out_dir)) dir.create(out_dir)
+
+######## Land use types comparison with LUH ########
+magpie_LU <- land(gdx, level = "cell")
+historical_LU <- readGDX(gdx, "f10_land")
+
+LU <- plotCorrHist2D(x = historical_LU, y = magpie_LU, xlab = "LUH2 \n (mio.ha)", ylab = "MAgPIE \n (mio.ha)", bins = 35,
+    folder = out_dir, file = "/validationLUcell", breaks = c(1,10,100), nrows=2, ncols=2, axisFont=10,
+                           axisTitleFont=11,TitleFontSize=12,legendTitleFont=10,legendTextFont=8,statFont=2,stat=FALSE, table=TRUE,tag="year")
+
+######## Crop types ("maiz","rice_pro","soybean","tece") comparison  with LUH ########
+magpie <- croparea(gdx, level = "cell", product_aggr = FALSE, water_aggr = TRUE)[, getYears(historical_LU), crops]
+historical1 <- dimOrder(readGDX(gdx, "fm_croparea")[, getYears(historical_LU), crops], perm = c(2, 1), dim = 3)
+historical1 <- dimSums(historical1,dim=3.2)
+
+
+Crops <- plotCorrHist2D(
+    x = historical1, y = magpie, xlab = "LUH2 \n (mio.ha)", ylab = "MAgPIE \n (mio.ha)", bins = 35,
+    folder = out_dir, file = "/validationCropCell", breaks = c(1,10,50), nrows=2, ncols=2, axisFont=10,
+                           axisTitleFont=11,TitleFontSize=12,legendTitleFont=10,legendTextFont=8,statFont=2,stat=FALSE, table=TRUE,tag="year")
+
+######## Crop types ("maiz","rice_pro","soybean","tece") comparison  with SPAM ########
+
+mapping <- readRDS(map_file)
+historical <- readSource(type="MAPSPAM",subtype = "physical")[,,crops]
+historical <- magpiesort(toolAggregate(historical,rel=mapping,from="cell",to="cluster"))/1e6
+historical <- dimSums(historical, dim=3.2)
+
+CropsSPAM <- plotCorrHist2D(
+    x = historical, y = magpie[,getYears(historical),], xlab = "MAPSPAM \n (mio.ha)", ylab = "MAgPIE \n (mio.ha)", bins = 35,
+    folder = out_dir, file = "/validationCropCellSPAM", breaks = c(1, 10, 50), nrows = 2, ncols = 2, axisFont=10,
+                           axisTitleFont=11,TitleFontSize=12,legendTitleFont=10,legendTextFont=8,statFont=2,stat=FALSE, table=TRUE,
+                           palette = "PiYG",tag="year"
+)
+
+######################################### Generates document ##############################################
+template <- c(
+    "\\documentclass[a4paper, portrait ]{article}",
+    "\\setlength{\\parindent}{0in}",
+    "\\usepackage{float}",
+    "\\usepackage[bookmarksopenlevel=section]{hyperref}",
+    "\\hypersetup{bookmarks=true,pdfauthor={PIK Landuse group}}",
+    "\\usepackage{graphicx}",
+    "\\usepackage{rotating}",
+    "\\usepackage[strings]{underscore}",
+    "\\usepackage[margin=2cm]{geometry}",
+    "\\usepackage{fancyhdr}",
+    "\\pagestyle{fancy}",
+    "\\begin{document}",
+    "<<echo=false>>=",
+    "options(width=90)",
+    "@"
+)
+
+############################## Plots LUH vs. magpie land use types ##########################################################
+x<-LU[[1]]
+y<-as.data.frame(LU[[2]])
+colnames(y)<-y[1,]
+rownames(y)<-NULL
+y<-(y[2:(nrow(y)),])
+y[,3:ncol(y)]<-round(as.numeric(unlist(y[,3:ncol(y)])),3)
+colnames(y)[2]<-"Year"
+colnames(y)[6] <- "MAEH*"
+
+tmpplot <- function(x,range){
+    plotLU<-function(x,range){
+    print(gridExtra::marrangeGrob(x[range], nrow = 3, ncol = 2, top = NULL))
+    }
+     a <- capture.output(plotLU(x,range))
+ }
+
+namesReport <- as.character(magpiesets::reportingnames(getNames(magpie_LU)))
+
+sw <- swopen(outfile = paste0(outputdir,"/CellValidation.pdf"), template = template)
+swlatex(sw, c("\\title{MAgPIE Validation on Cellular level}",  "\\maketitle", "\\tableofcontents"))
+swlatex(sw, "\\section{Land use between 1995-2020}")
+
+b<-0
+
+for(i in 1:length(namesReport)){
+swlatex(sw, paste0("\\subsection{",namesReport[i],"}"))
+a<-b+1
+b<-a+length(getYears(historical_LU))-1
+swfigure(sw, tmpplot,x=x,c(a:b))
+swtable(sw, collapseNames(as.magpie(y[a:b, ])), table.placement = "H", caption.placement = "top", 
+        transpose = FALSE, caption = paste0("Compilation of error measurements for ", namesReport[i] , " data comparing MAgPIE outputs \\& the LUH2 set at cluster level"), vert.lines = 0, align = "c")
+}
+swlatex(sw, "\\newpage")
+
+############################## Plots LUH vs. magpie crop types ##########################################################
+swlatex(sw, "\\section{Crop types 1995-2015 LUH}")
+
+namesReport1 <- as.character(magpiesets::reportingnames(getNames(magpie)))
+x1 <- Crops[[1]]
+y1 <- as.data.frame(Crops[[2]])
+colnames(y1) <- y1[1, ]
+rownames(y1) <- NULL
+y1 <- (y1[2:(nrow(y1)), ])
+y1[, 3:ncol(y1)] <- round(as.numeric(unlist(y1[, 3:ncol(y1)])), 3)
+colnames(y1)[2] <- "Year"
+colnames(y1)[6] <- "MAEH*"
+
+tmpplot1 <- function(x1, range) {
+    plotLU1 <- function(x1, range) {
+        print(gridExtra::marrangeGrob(x1[range], nrow = 3, ncol = 2, top = NULL))
+    }
+    a <- capture.output(plotLU1(x1, range))
+}
+
+b1 <- 0
+for (i in 1:length(namesReport1)) {
+    swlatex(sw, paste0("\\subsection{", namesReport1[i], "}"))
+    a1 <- b1 + 1
+    b1 <- a1 + length(getYears(historical1)) - 1
+    swfigure(sw, tmpplot1, x = x1, c(a1:b1))
+    swtable(sw, collapseNames(as.magpie(y1[a1:b1, ])),
+        table.placement = "H", caption.placement = "top",
+        transpose = FALSE, caption = paste0("Compilation of error measurements for ", namesReport[i] , " data comparing MAgPIE outputs \\& the LUH2 set at cluster level"), vert.lines = 0, align = "c"
+    )
+}
+
+
+############################## Plots MAPSPAM vs. magpie crop types ##########################################################
+swlatex(sw, "\\newpage")
+swlatex(sw, "\\section{Crop types 2000-2010 - MAPSPAM and LUH comparison}")
+
+x2 <-CropsSPAM[[1]]
+names(x2) <- paste0(names(x2),"MAP")
+x3<-Crops[[1]][names(CropsSPAM[[1]])]
+y2 <- as.data.frame(CropsSPAM[[2]])
+colnames(y2) <- y2[1, ]
+rownames(y2) <- NULL
+y2 <- (y2[2:(nrow(y2)), ])
+y2[, 3:ncol(y2)] <- round(as.numeric(unlist(y2[, 3:ncol(y2)])), 3)
+colnames(y2)[2] <- "Year"
+colnames(y2)[6] <- "MAEH*"
+
+tmpplot2 <- function(x2,x3,range) {
+    plotLU2 <- function(x2,x3,range) {
+        print(gridExtra::marrangeGrob(c(x2[range],x3[range]), nrow = 3, ncol = 2, top = NULL))
+    }
+    a <- capture.output(plotLU2(x2,x3, range))
+}
+
+b2 <- 0
+for (i in 1:length(namesReport1)) {
+    swlatex(sw, paste0("\\subsection{", namesReport1[i], "}"))
+    swlatex(sw, "\\qquad \\textbf{MAPSPAM} \\qquad \\qquad \\qquad \\qquad \\qquad \\qquad \\qquad \\qquad \\qquad \\qquad \\textbf{LUH2}")
+    a2 <- b2 + 1
+    b2 <- a2 + length(getYears(historical)) - 1
+    swfigure(sw, tmpplot2, x2 = x2, x3 = x3, range= c(a2:b2))
+    swtable(sw, collapseNames(as.magpie(y2[a2:b2, ])),
+        table.placement = "H", caption.placement = "top",
+        transpose = FALSE, caption = paste0("Compilation of error measurements for ", namesReport[i] , " data comparing MAgPIE outputs \\& the LUH2 set at cluster level"), vert.lines = 0, align = "c")
+    swlatex(sw, "\\newpage")
+    
+}
+
+########## Details about MAEH ##########
+swlatex(sw, "\\newpage")
+swlatex(sw, "*MAEH Stands for Mean Absolute Error per hectar. It is calculated as $MAE=\\sum(|error|)/area$ . Here, error represents the difference between LUH/MAPSPAM values and MAgPIE Outputs.")
+swlatex(sw, " Area is the overal world area for the specific land use or crop type in time t as reported by LUH2 or MAPSPAM")
+
+####### Closes the pdf document
+swclose(sw, clean_output = FALSE, engine = "knitr")
+
+###########################################################################################################################
+
+##### Erase latex style and output files
+
+fileErase <- c(Sys.glob(file.path(outputdir, "/CellValidation.*")),
+               Sys.glob(file.path(outputdir, "/Sweave.sty")),
+               Sys.glob(file.path(outputdir, "figure/*")),
+               Sys.glob(file.path(outputdir, "figure")))
+fileErase <- fileErase[!(fileErase %in% paste0(outputdir, "//CellValidation.pdf"))]
+
+
+unlink(fileErase,recursive = TRUE)